refactor: 뒤로가기 Topbar 공용 컴포넌트로 분리
